Put save path function in path generator
