+++ mirrulations-core/src/mirrcore/path_generator.py
@@ -154,3 +154,24 @@ class PathGenerator:
                 print(f"fileFormats did not exist for attachment ID: {id}")
 
         return attachments
+    
+    @staticmethod
+    def make_attachment_save_path(path):
+        '''
+        This method takes a complete path to a pdf and makes
+        the save path based on paramters in that path.
+        Parameters
+        ----------
+        path : str
+            the complete file path for the attachment that is being extracted
+            ex. /path/to/pdf/attachment_1.pdf
+        '''
+        if "comments" in path:
+            return path.replace('binary', 'text') \
+            .replace('comments_attachments', 'comments_extracted_text/pdfminer') \
+            .replace('.pdf', '_extracted.txt') \
+        
+        return path.replace('binary', 'text') \
+            .replace('documents_attachments', 'documents_extracted_text/pdfminer') \
+            .replace('.pdf', '_extracted.txt') \
+
